+++ src/core/config.rs
@@ -4,10 +4,23 @@ use std::path::{Path, PathBuf};
 use std::{env, fs};
 
 #[derive(Debug, Clone)]
+pub enum DbType {
+    Sqlite,
+}
+
+#[derive(Debug, Clone)]
+pub enum IntegrationType {
+    Redmine,
+}
+
+
+#[derive(Debug, Clone)]
 pub struct Config {
     pub app_share_path: PathBuf,
     pub redmine_url: Option<String>,
     pub redmine_token: Option<String>,
+    pub db_type: DbType,
+    pub int_type: IntegrationType,
 }
 
 #[derive(Debug)]
@@ -15,6 +28,8 @@ struct ConfigFile {
     pub app_share_path: Option<PathBuf>,
     pub redmine_url: Option<String>,
     pub redmine_token: Option<String>,
+    pub db_type: Option<DbType>,
+    pub int_type: Option<IntegrationType>,
 }
 
 impl Config {
@@ -55,17 +70,29 @@ impl Config {
                 let redmine_token = ini_file_props
                     .get("redmine_token")
                     .map(|value| value.to_string());
+                let db_type = ini_file_props
+                    .get("general_db")
+                    .map(|value| DbType::from_string(value))
+                    .flatten();
+                let int_type = ini_file_props
+                    .get("general_integration")
+                    .map(|value| IntegrationType::from_string(value))
+                    .flatten();
 
                 ConfigFile {
                     app_share_path,
                     redmine_url,
                     redmine_token,
+                    db_type,
+                    int_type,
                 }
             }
             Err(_) => ConfigFile {
                 app_share_path: None,
                 redmine_url: None,
                 redmine_token: None,
+                db_type: None,
+                int_type: None,
             },
         }
     }
@@ -76,6 +103,8 @@ impl Config {
             app_share_path,
             redmine_url: None,
             redmine_token: None,
+            db_type: DbType::Sqlite,
+            int_type: IntegrationType::Redmine,
         }
     }
 
@@ -87,6 +116,8 @@ impl Config {
                 .unwrap_or(default.app_share_path.clone()),
             redmine_url: from_ini.redmine_url.clone().or(default.redmine_url.clone()),
             redmine_token: from_ini.redmine_token.clone().or(default.redmine_token.clone()),
+            db_type: from_ini.db_type.clone().unwrap_or(default.db_type.clone()),
+            int_type: from_ini.int_type.clone().unwrap_or(default.int_type.clone()),
         }
     }
 
@@ -94,6 +125,24 @@ impl Config {
         if fs::metadata(&share_path).is_err() {
             fs::create_dir_all(&share_path)
                 .expect("'mytime' configuration folder has not been created");
+        }
+    }
+}
+
+impl DbType {
+    pub fn from_string(value: &str) -> Option<DbType> {
+        match value.to_lowercase().as_str() {
+            "sqlite" => Some(DbType::Sqlite),
+            _ => None,
+        }
+    }
+}
+
+impl IntegrationType {
+    pub fn from_string(value: &str) -> Option<IntegrationType> {
+        match value.to_lowercase().as_str() {
+            "redmine" => Some(IntegrationType::Redmine),
+            _ => None,
         }
     }
 }