+++ src/integrations/traits.rs
@@ -3,5 +3,5 @@ use crate::core::task::Task;
 use crate::core::config::Config;
 
 pub trait Integration {
-    fn report_task<'a>(config: &'a Config, task: &Task) -> Result<(), Error>;
+    fn report_task<'a>(&self, config: &'a Config, task: &Task) -> Result<(), Error>;
 }