+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ### Added
 
 - Show task preview after modify it
+- Allow set db and integration through config file
 
 ## [0.1.2] - 2023-06-03